Refactors CLM nav for permission access
Migrates the CLM navigation items from role-based access control to a more granular permission-based system.

Introduces a dedicated `Permission` enum for the CLM module, ensuring frontend authorization aligns with backend definitions. Deprecates the `forRole` property in nav item types, promoting the use of `forPermission` for more precise access management.
